0.8 законченное приложение
