+++ .idea/encodings.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/выгрузка_задачи.txt" charset="windows-1251" />
+    <file url="file://$PROJECT_DIR$/выгрузка_пользователей.txt" charset="windows-1251" />
+  </component>
+</project>
